requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.6.2.post1 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-awscli-v1==2.2.211 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-kubectl-v20==2.1.3 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-node-proxy-agent-v6==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-aws-lambda-python-alpha==2.167.0a0 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-cloud-assembly-schema==38.0.1 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-lib==2.167.0 ; python_version >= "3.12" and python_version < "4.0"
cattrs==24.1.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
constructs==10.4.2 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.5 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-resources==6.4.5 ; python_version >= "3.12" and python_version < "4.0"
jsii==1.105.0 ; python_version >= "3.12" and python_version < "4.0"
mangum==0.19.0 ; python_version >= "3.12" and python_version < "4.0"
publication==0.0.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.23.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.41.2 ; python_version >= "3.12" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.32.0 ; python_version >= "3.12" and python_version < "4.0"
